backend/requirements.txt
fastapi
uvicorn
websockets
requests
reportlab
python-dotenv
pydantic
